+++ src/index.ts
@@ -10,7 +10,7 @@ export { ButtonGroup } from './components/Button'
 export { default as Chat } from './components/Chat'
 export { ChatItem } from './components/Chat'
 export { ChatMessage } from './components/Chat'
-export { default as ContextualMenu } from './components/ContextualMenu'
+export { default as ContextMenu } from './components/ContextMenu'
 export { default as Divider } from './components/Divider'
 export { default as Grid } from './components/Grid'
 export { default as Header } from './components/Header'
