+++ src/themes/teams/componentStyles.ts
@@ -13,7 +13,7 @@ export { default as Chat } from './components/Chat/chatStyles'
 export { default as ChatItem } from './components/Chat/chatItemStyles'
 export { default as ChatMessage } from './components/Chat/chatMessageStyles'
 
-export { default as ContextualMenu } from './components/ContextualMenu/contextualMenuStyles'
+export { default as ContextMenu } from './components/ContextMenu/contextMenuStyles'
 
 export { default as Divider } from './components/Divider/dividerStyles'
