+++ src/themes/teams/componentVariables.ts
@@ -13,7 +13,7 @@ export { default as ChatItem } from './components/Chat/chatItemVariables'
 
 export { default as ChatMessage } from './components/Chat/chatMessageVariables'
 
-export { default as ContextualMenu } from './components/ContextualMenu/contextualMenuVariables'
+export { default as ContextMenu } from './components/ContextMenu/contextMenuVariables'
 
 export { default as Divider } from './components/Divider/dividerVariables'
